random walk on routing graph
